requirements.txt
langchain
langchain-openai
langchain-community
langchain-core
langchain-experimental
langchain-chroma
streamlit
yfinance
chromadb
python-dotenv
tavily-python
SQLAlchemy